+++ app/posts/[id]/page.tsx
@@ -5,6 +5,7 @@ import { useParams } from "next/navigation";
 import { PostContext } from "../../../context/PostsContext";
 import { Blog } from "../../../context/PostsContext";
 import NoDataPlaceHolder from "../../../components/NoDataPlaceholder";
+import SingleBlogItem from "../../../components/SingleBlogItem";
 
 export default function Post() {
   const { blogs } = useContext(PostContext);
@@ -16,9 +17,8 @@ export default function Post() {
   return (
     <>
       {selectedBlog ? (
-        <div>
-          <h1>{selectedBlog.title}</h1>
-          <p>{selectedBlog.body}</p>
+        <div className="flex flex-col items-center justify-center min-h-screen">
+          <SingleBlogItem blog={selectedBlog} />
         </div>
       ) : (
         <NoDataPlaceHolder />
