+++ components/DataTable.tsx
@@ -5,7 +5,7 @@ import { PostContext } from "../context/PostsContext";
 import PageErrorState from "./PageErrorState";
 import PageLoadingState from "./PageLoadingState";
 import NoDataPlaceHolder from "./NoDataPlaceholder";
-import BlogItem from "./BlogItem";
+import AllBlogItem from "./AllBlogItem";
 
 export default function DataTable() {
   const { blogs, loading, error } = useContext(PostContext);
@@ -13,7 +13,6 @@ export default function DataTable() {
   if (loading) {
     return <PageLoadingState />;
   }
-
   if (error) {
     return <PageErrorState />;
   }
@@ -23,9 +22,9 @@ export default function DataTable() {
       {Object.keys(blogs).length === 0 ? (
         <NoDataPlaceHolder />
       ) : (
-        <ul className="divide-y divide-gray-200 max-w-xl mx-auto">
+        <ul className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-8 max-w-6xl mx-auto">
           {blogs.map((blog) => (
-            <BlogItem key={blog.id} blog={blog} />
+            <AllBlogItem key={blog.id} blog={blog} />
           ))}
         </ul>
       )}
